faceoff/faceoff/requirements.txt
numpy==1.22.1
streamlit==1.4.0
opencv-contrib-python==4.5.5.62
Pillow==9.0.0
